Move permission related utils to model
